bin/backend_side_by_django_on_aws_version_4/django_with_PostgreSQL_app/serializers.py
from .models import data_logger_model
from rest_framework import serializers

class data_logger_serializer(serializers.ModelSerializer):
    class Meta:
        model = data_logger_model
        fields = ['id','sensor', 'water_level', 'longitude', 'latitude', 'location', 'date_and_time']